# storage/api-docs/api-docs.json
{
    "swagger": "2.0",
    "info": {
        "title": " App",
        "version": "v1.0.0"
    },
    "paths": {
...
        "/api/user/bindEmail": {
            "get": {
                "tags": [
                    "User"
                ],
                "summary": "\u83b7\u5f97\u90ae\u7bb1\u9a8c\u8bc1\u7801",
                "description": "\u90ae\u7bb1\u9a8c\u8bc1\u7801",
                "operationId": "",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "Token",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "Content-Type",
                        "in": "header",
                        "required": true,
                        "type": "string",
                        "default": "application/json"
                    },
                    {
                        "name": "email",
                        "in": "path",
                        "description": "\u90ae\u7bb1",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "default": {
                        "description": "\u64cd\u4f5c\u6210\u529f"
                    }
                }
            },
            "post": {
                "tags": [
                    "User"
                ],
                "summary": "\u7ed1\u5b9a\u90ae\u7bb1",
                "description": "\u7ed1\u5b9a\u90ae\u7bb1",
                "operationId": "",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "Token",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "Content-Type",
                        "in": "header",
                        "required": true,
                        "type": "string",
                        "default": "application/json"
                    },
                    {
                        "name": "code",
                        "in": "formData",
                        "description": "\u90ae\u7bb1\u9a8c\u8bc1\u7801",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "email",
                        "in": "formData",
                        "description": "\u90ae\u7bb1\u5730\u5740",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "key",
                        "in": "formData",
                        "description": "\u4e0a\u4e00\u6b65\u83b7\u53d6\u90ae\u7bb1\u9a8c\u8bc1\u7801\u8fd4\u56de\u7684key",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "default": {
                        "description": "\u64cd\u4f5c\u6210\u529f"
                    }
                }
            }
        },
        "/api/user": {
            "put": {
                "tags": [
                    "User"
                ],
                "summary": "\u66f4\u65b0\u7528\u6237\u4fe1\u606f",
                "description": "\u66f4\u65b0",
                "operationId": "",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "Token",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "Content-Type",
                        "in": "header",
                        "required": true,
                        "type": "string",
                        "default": "application/json"
                    },
                    {
                        "name": "name",
                        "in": "formData",
                        "description": "\u65b0\u7684\u7528\u6237\u6635\u79f0",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "avatar",
                        "in": "formData",
                        "description": "\u7528\u6237\u5934\u50cf",
                        "required": false,
                        "type": "number"
                    }
                ],
                "responses": {
                    "default": {
                        "description": "\u64cd\u4f5c\u6210\u529f"
                    }
                }
            }
        },
        "/api/smsCode": {
            "get": {
                "tags": [
                    "Auth"
                ],
                "summary": "\u83b7\u53d6\u624b\u673a\u6ce8\u518c\u9a8c\u8bc1\u7801",
                "description": "\u83b7\u53d6\u9a8c\u8bc1\u7801",
                "operationId": "",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "mobile",
                        "in": "query",
                        "description": "\u624b\u673a\u53f7",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "default": {
                        "description": "\u64cd\u4f5c\u6210\u529f"
                    }
                }
            }
        },
        "/api/register": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "\u901a\u8fc7\u624b\u673a\u9a8c\u8bc1\u7801\u6ce8\u518c\u7528\u6237",
                "description": "\u6ce8\u518c",
                "operationId": "",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "verification_key",
                        "in": "formData",
                        "description": "\u83b7\u53d6\u9a8c\u8bc1\u7801\u6b65\u9aa4\u4e2d\u8fd4\u56de\u7684key",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "verification_code",
                        "in": "formData",
                        "description": "\u624b\u673a\u9a8c\u8bc1\u7801",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "mobile",
                        "in": "formData",
                        "description": "\u624b\u673a\u53f7",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "password",
                        "in": "formData",
                        "description": "\u7528\u6237\u5bc6\u7801",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "default": {
                        "description": "\u64cd\u4f5c\u6210\u529f"
                    }
                }
            }
        }
    },
...
}